+++ iohub/ngff/models.py
